+++ age/src/ssh/recipient.rs
@@ -21,7 +21,10 @@ use super::{
     read_ssh, ssh_tag, SSH_ED25519_KEY_PREFIX, SSH_ED25519_RECIPIENT_KEY_LABEL,
     SSH_ED25519_RECIPIENT_TAG, SSH_RSA_KEY_PREFIX, SSH_RSA_OAEP_LABEL, SSH_RSA_RECIPIENT_TAG,
 };
-use crate::util::read::{encoded_str, str_while_encoded};
+use crate::{
+    error::EncryptError,
+    util::read::{encoded_str, str_while_encoded},
+};
 
 /// A key that can be used to encrypt a file to a recipient.
 #[derive(Clone, Debug)]
@@ -70,7 +73,7 @@ impl fmt::Display for Recipient {
 }
 
 impl crate::Recipient for Recipient {
-    fn wrap_file_key(&self, file_key: &FileKey) -> Stanza {
+    fn wrap_file_key(&self, file_key: &FileKey) -> Result<Stanza, EncryptError> {
         match self {
             Recipient::SshRsa(ssh_key, pk) => {
                 let mut rng = OsRng;
@@ -86,11 +89,11 @@ impl crate::Recipient for Recipient {
                 let encoded_tag =
                     base64::encode_config(&ssh_tag(&ssh_key), base64::STANDARD_NO_PAD);
 
-                Stanza {
+                Ok(Stanza {
                     tag: SSH_RSA_RECIPIENT_TAG.to_owned(),
                     args: vec![encoded_tag],
                     body: encrypted_file_key,
-                }
+                })
             }
             Recipient::SshEd25519(ssh_key, ed25519_pk) => {
                 let pk: X25519PublicKey = ed25519_pk.to_montgomery().to_bytes().into();
@@ -119,11 +122,11 @@ impl crate::Recipient for Recipient {
                     base64::encode_config(&ssh_tag(&ssh_key), base64::STANDARD_NO_PAD);
                 let encoded_epk = base64::encode_config(epk.as_bytes(), base64::STANDARD_NO_PAD);
 
-                Stanza {
+                Ok(Stanza {
                     tag: SSH_ED25519_RECIPIENT_TAG.to_owned(),
                     args: vec![encoded_tag, encoded_epk],
                     body: encrypted_file_key,
-                }
+                })
             }
         }
     }